+++ packages/i18nify-js/src/modules/phoneNumber/data/phoneRegexMapper.ts
@@ -1,4 +1,4 @@
-import { CountryCodeType } from '../../shared/types';
+import { CountryCodeType } from '../../types/geo';
 
 export const PHONE_REGEX_MAPPER: { [key in CountryCodeType]: RegExp } = {
   IN: /^(?:(?:\+|0{0,2})91\s*[-]?\s*|[0]?)?[6789]\d{9}$/,
@@ -176,76 +176,76 @@ export const PHONE_REGEX_MAPPER: { [key in CountryCodeType]: RegExp } = {
   GA: /^(?:\+241)?(?:\d{6,7})$/,
   GD: /^(?:\+1-473)?(?:\d{7})$/,
   GE: /^(?:\+995)?(?:\d{8,9})$/,
-  GU: /^(?:\+1-671)?\d{7}$/, 
-  KN: /^(?:\+1-869)?\d{7}$/, 
-  LC: /^(?:\+1-758)?\d{7}$/, 
-  MP: /^(?:\+1-670)?\d{7}$/, 
-  MS: /^(?:\+1-664)?\d{7}$/, 
-  SX: /^(?:\+1-721)?\d{7}$/, 
-  TC: /^(?:\+1-649)?\d{7}$/, 
-  VG: /^(?:\+1-284)?\d{7}$/, 
-  VI: /^(?:\+1-340)?\d{7}$/, 
-  GR: /^(?:\+30)?\d{10}$/, 
-  IT: /^(?:\+39)?\d{9,10}$/, 
-  VA: /^(?:\+379)?\d{8,12}$/, 
-  GG: /^(?:\+44-1481)?\d{6}$/, 
-  IM: /^(?:\+44-1624)?\d{6}$/, 
-  JE: /^(?:\+44-1534)?\d{6}$/, 
-  SJ: /^\\+47\\d{8}$/, 
-  JP: /^(?:\+81)?\d{10}$/, 
-  KR: /^(?:\+82)?\d{8,11}$/, 
-  AF: /^(?:\+93)?\d{9}$/, 
-  IR: /^(?:\+98)?\d{10}$/, 
-  LY: /^(?:\+218)?\d{9}$/, 
-  MR: /^(?:\+222)?\d{8}$/, 
-  ML: /^(?:\+223)?\d{8}$/, 
-  GN: /^(?:\+224)?\d{8}$/, 
-  NE: /^(?:\+227)?\d{8}$/, 
-  TD: /^(?:\+235)?\d{8}$/, 
-  ST: /^(?:\+239)?\d{7}$/, 
-  GQ: /^(?:\+240)?\d{9}$/, 
-  GW: /^(?:\+245)?\d{7}$/, 
-  IO: /^\\+246\\d{7}$/, 
-  AC: /^\\+247\\d{4}$/, 
-  SD: /^(?:\+249)?\d{9}$/, 
-  RE: /^(?:\+262)?\d{9}$/, 
-  YT: /^(?:\+262)?\d{9}$/, 
-  KM: /^(?:\+269)?\d{7}$/, 
-  SH: /^\\+290\\d{4,5}$/, 
-  TA: /^\\+290\\d{4}$/, 
-  GL: /^(?:\+299)?\d{6}$/, 
-  IE: /^(?:\+353)?\d{7,10}$/, 
-  IS: /^(?:\+354)?\d{7}$/, 
-  MT: /^(?:\+356)?\d{8}$/, 
-  AX: /^\\+358\\d{5,8}$/, 
-  AD: /^(?:\+376)?\d{6}$/, 
-  MC: /^(?:\+377)?\d{8,9}$/, 
-  XK: /^\\+383\\d{8}$/, 
-  LI: /^(?:\+423)?\d{7}$/, 
-  PM: /^(?:\+508)?\d{6}$/, 
-  GP: /^(?:\+590)?\d{9}$/, 
-  MF: /^(?:\+590)?\d{9}$/, 
-  GF: /^(?:\+594)?\d{9}$/, 
-  MQ: /^(?:\+596)?\d{9}$/, 
-  NF: /^(?:\+672)?\d{6}$/, 
-  NR: /^(?:\+674)?\d{7}$/, 
-  TO: /^(?:\+676)?\d{5}$/, 
-  SB: /^(?:\+677)?\d{5}$/, 
-  VU: /^(?:\+678)?\d{7}$/, 
-  PW: /^(?:\+680)?\d{7}$/, 
-  WF: /^(?:\+681)?\d{6}$/, 
-  NU: /^(?:\+683)?\d{4}$/, 
-  WS: /^(?:\+685)?\d{5}$/, 
-  KI: /^(?:\+686)?\d{5}$/, 
-  NC: /^(?:\+687)?\d{6}$/, 
-  TV: /^(?:\+688)?\d{5}$/, 
-  PF: /^(?:\+689)?\d{6}$/, 
-  TK: /^\\+690\\d{4}$/, 
-  MH: /^(?:\+692)?\d{7}$/, 
-  KP: /^\\+850\\d{6,9}$/, 
-  LB: /^(?:\+961)?\d{8}$/, 
-  JO: /^(?:\+962)?\d{9}$/, 
-  SY: /^(?:\+963)?\d{9}$/, 
-  IQ: /^(?:\+964)?\d{10}$/, 
-  TM: /^(?:\+993)?\d{8}$/, 
+  GU: /^(?:\+1-671)?\d{7}$/,
+  KN: /^(?:\+1-869)?\d{7}$/,
+  LC: /^(?:\+1-758)?\d{7}$/,
+  MP: /^(?:\+1-670)?\d{7}$/,
+  MS: /^(?:\+1-664)?\d{7}$/,
+  SX: /^(?:\+1-721)?\d{7}$/,
+  TC: /^(?:\+1-649)?\d{7}$/,
+  VG: /^(?:\+1-284)?\d{7}$/,
+  VI: /^(?:\+1-340)?\d{7}$/,
+  GR: /^(?:\+30)?\d{10}$/,
+  IT: /^(?:\+39)?\d{9,10}$/,
+  VA: /^(?:\+379)?\d{8,12}$/,
+  GG: /^(?:\+44-1481)?\d{6}$/,
+  IM: /^(?:\+44-1624)?\d{6}$/,
+  JE: /^(?:\+44-1534)?\d{6}$/,
+  SJ: /^\\+47\\d{8}$/,
+  JP: /^(?:\+81)?\d{10}$/,
+  KR: /^(?:\+82)?\d{8,11}$/,
+  AF: /^(?:\+93)?\d{9}$/,
+  IR: /^(?:\+98)?\d{10}$/,
+  LY: /^(?:\+218)?\d{9}$/,
+  MR: /^(?:\+222)?\d{8}$/,
+  ML: /^(?:\+223)?\d{8}$/,
+  GN: /^(?:\+224)?\d{8}$/,
+  NE: /^(?:\+227)?\d{8}$/,
+  TD: /^(?:\+235)?\d{8}$/,
+  ST: /^(?:\+239)?\d{7}$/,
+  GQ: /^(?:\+240)?\d{9}$/,
+  GW: /^(?:\+245)?\d{7}$/,
+  IO: /^\\+246\\d{7}$/,
+  AC: /^\\+247\\d{4}$/,
+  SD: /^(?:\+249)?\d{9}$/,
+  RE: /^(?:\+262)?\d{9}$/,
+  YT: /^(?:\+262)?\d{9}$/,
+  KM: /^(?:\+269)?\d{7}$/,
+  SH: /^\\+290\\d{4,5}$/,
+  TA: /^\\+290\\d{4}$/,
+  GL: /^(?:\+299)?\d{6}$/,
+  IE: /^(?:\+353)?\d{7,10}$/,
+  IS: /^(?:\+354)?\d{7}$/,
+  MT: /^(?:\+356)?\d{8}$/,
+  AX: /^\\+358\\d{5,8}$/,
+  AD: /^(?:\+376)?\d{6}$/,
+  MC: /^(?:\+377)?\d{8,9}$/,
+  XK: /^\\+383\\d{8}$/,
+  LI: /^(?:\+423)?\d{7}$/,
+  PM: /^(?:\+508)?\d{6}$/,
+  GP: /^(?:\+590)?\d{9}$/,
+  MF: /^(?:\+590)?\d{9}$/,
+  GF: /^(?:\+594)?\d{9}$/,
+  MQ: /^(?:\+596)?\d{9}$/,
+  NF: /^(?:\+672)?\d{6}$/,
+  NR: /^(?:\+674)?\d{7}$/,
+  TO: /^(?:\+676)?\d{5}$/,
+  SB: /^(?:\+677)?\d{5}$/,
+  VU: /^(?:\+678)?\d{7}$/,
+  PW: /^(?:\+680)?\d{7}$/,
+  WF: /^(?:\+681)?\d{6}$/,
+  NU: /^(?:\+683)?\d{4}$/,
+  WS: /^(?:\+685)?\d{5}$/,
+  KI: /^(?:\+686)?\d{5}$/,
+  NC: /^(?:\+687)?\d{6}$/,
+  TV: /^(?:\+688)?\d{5}$/,
+  PF: /^(?:\+689)?\d{6}$/,
+  TK: /^\\+690\\d{4}$/,
+  MH: /^(?:\+692)?\d{7}$/,
+  KP: /^\\+850\\d{6,9}$/,
+  LB: /^(?:\+961)?\d{8}$/,
+  JO: /^(?:\+962)?\d{9}$/,
+  SY: /^(?:\+963)?\d{9}$/,
+  IQ: /^(?:\+964)?\d{10}$/,
+  TM: /^(?:\+993)?\d{8}$/,
 };
