+++ packages/i18nify-js/src/modules/phoneNumber/getDialCodes.ts
@@ -1,5 +1,5 @@
 import { withErrorBoundary } from '../../common/errorBoundary';
-import { CountryCodeType } from '../shared/types';
+import { CountryCodeType } from '../types/geo';
 import { DIAL_CODE_MAPPER } from './data/dialCodeMapper';
 
 /**
