+++ packages/i18nify-js/src/modules/phoneNumber/isValidPhoneNumber.ts
@@ -1,7 +1,7 @@
 import { PHONE_REGEX_MAPPER } from './data/phoneRegexMapper';
 import { withErrorBoundary } from '../../common/errorBoundary';
 import { detectCountryAndDialCodeFromPhone, cleanPhoneNumber } from './utils';
-import { CountryCodeType } from '../shared/types';
+import { CountryCodeType } from '../types/geo';
 
 // Validates whether a given phone number is valid based on the provided country code or auto-detects the country code and checks if the number matches the defined regex pattern for that country.
 const isValidPhoneNumber = (
@@ -12,10 +12,11 @@ const isValidPhoneNumber = (
   const cleanedPhoneNumber = cleanPhoneNumber(phoneNumber.toString());
 
   // Detect or validate the country code
-  countryCode =
-    (countryCode && countryCode in PHONE_REGEX_MAPPER
+  countryCode = (
+    countryCode && countryCode in PHONE_REGEX_MAPPER
       ? countryCode
-      : detectCountryAndDialCodeFromPhone(cleanedPhoneNumber).countryCode) as CountryCodeType;
+      : detectCountryAndDialCodeFromPhone(cleanedPhoneNumber).countryCode
+  ) as CountryCodeType;
 
   // Return false if phoneNumber is empty
   if (!phoneNumber) return false;
