+++ packages/i18nify-js/src/modules/phoneNumber/parsePhoneNumber.ts
@@ -1,5 +1,5 @@
 import { withErrorBoundary } from '../../common/errorBoundary';
-import { CountryCodeType } from '../shared/types';
+import { CountryCodeType } from '../types/geo';
 import { PHONE_FORMATTER_MAPPER } from './data/phoneFormatterMapper';
 import formatPhoneNumber from './formatPhoneNumber';
 import { detectCountryAndDialCodeFromPhone, cleanPhoneNumber } from './utils';
@@ -13,7 +13,10 @@ interface PhoneInfo {
 }
 
 // Parses a given phone number, identifies its country code (if not provided), and returns an object with details including the country code, formatted phone number, dial code, and format template.
-const parsePhoneNumber = (phoneNumber: string, country?: CountryCodeType): PhoneInfo => {
+const parsePhoneNumber = (
+  phoneNumber: string,
+  country?: CountryCodeType,
+): PhoneInfo => {
   // Throw errors if phoneNumber is invalid
   if (!phoneNumber) throw new Error('Parameter `phoneNumber` is invalid!');
 
@@ -23,10 +26,11 @@ const parsePhoneNumber = (phoneNumber: string, country?: CountryCodeType): Phone
 
   const countryData = detectCountryAndDialCodeFromPhone(phoneNumber);
   // Detect or validate the country code
-  const countryCode =
-    (country && country in PHONE_FORMATTER_MAPPER
+  const countryCode = (
+    country && country in PHONE_FORMATTER_MAPPER
       ? country
-      : countryData.countryCode) as CountryCodeType;
+      : countryData.countryCode
+  ) as CountryCodeType;
 
   const dialCode = countryData.dialCode;
   // Format the phone number using the detected/validated country code
