+++ packages/i18nify-js/src/modules/phoneNumber/utils.ts
@@ -1,4 +1,4 @@
-import { CountryCodeType } from '../shared/types';
+import { CountryCodeType } from '../types/geo';
 import { DIAL_CODE_MAPPER } from './data/dialCodeMapper';
 import { PHONE_REGEX_MAPPER } from './data/phoneRegexMapper';
