--- packages/i18nify-js/src/modules/shared/index.ts
@@ -1,1 +1,0 @@
-export * from './types';
